[ Api - Done ] - Seeder cambiado y añadido la nueva api que busca el libro
